refactor: PickUpSubscribe 엔티티 제거
* PickUpSubscribe 엔티티 속성을 PickUpSubscription 엔티티로 흡수
