+++ BE/src/main/java/ifive/idrop/entity/PickUpSubscription.java
@@ -5,6 +5,7 @@ import ifive.idrop.child.domain.Child;
 import ifive.idrop.common.exception.BusinessException;
 import ifive.idrop.common.exception.ErrorCode;
 import ifive.idrop.driver.domain.Driver;
+import ifive.idrop.entity.enums.PickUpStatus;
 import ifive.idrop.parent.domain.Parent;
 import ifive.idrop.parent.dto.SubscribeRequest;
 import ifive.idrop.pickup.domain.PickUp;
@@ -16,6 +17,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.RequiredArgsConstructor;
 
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -25,10 +27,14 @@ import java.util.List;
 @AllArgsConstructor
 @Getter
 public class PickUpSubscription {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "pickup_info_id")
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "pick_up_subscription_id")
     private Long id;
+    @Enumerated(EnumType.STRING)
+    private PickUpStatus status;
+    private LocalDateTime requestDate;
+    private LocalDateTime modifiedDate;
+    private LocalDateTime expiredDate;
 
     @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "pickup_location_id")
@@ -44,16 +50,8 @@ public class PickUpSubscription {
 
     private String schedule;
 
-    @OneToOne(fetch = FetchType.LAZY, mappedBy = "pickUpSubscription")
-    private PickUpSubscribe pickUpSubscribe;
-
     @OneToMany(mappedBy = "pickUpSubscription", cascade = CascadeType.ALL)
     private List<PickUpSchedule> pickUpScheduleList = new ArrayList<>();
-
-    public void updatePickUpSubscribe(PickUpSubscribe pickUpSubscribe) {
-        this.pickUpSubscribe = pickUpSubscribe;
-        pickUpSubscribe.setPickUpSubscription(this);
-    }
 
     public void updatePickUpLocation(PickUpLocation location) {
         this.pickUpLocation = location;
@@ -65,5 +63,18 @@ public class PickUpSubscription {
 
     public Parent getParent() {
         return this.child.getParent();
+    }
+
+    public PickUpStatus modify(PickUpStatus newStatus) {
+        this.status = newStatus;
+        //상태가 변경된 시간
+        this.modifiedDate = LocalDateTime.now();
+
+        if (this.status.equals(PickUpStatus.ACCEPT)) {
+            //modifiedDate로부터 29일 후 자정
+            this.expiredDate = this.modifiedDate.plusDays(29)
+                    .withHour(0).withMinute(0).withSecond(0).withNano(0);
+        }
+        return this.status;
     }
 }
